repos: Validate integrity of check-out against sha256sum
If a repo comes with a sha256sum field set, generate an archive from the
checked out revision and hash that using SHA-256. This allows to
identify potential supply chain attacks, e.g. when someone took over the
original repository or a used mirror and injected a modified version of
it. Various attacks are possible in that case, not only because of
weaknesses of the SHA-1 hashing mechanism that is used by default in git
and hg today.

Special care has to be taken for git because it allows to skip files on
export using export-ignore in the .gitattributes file. As that file can
be skipped itself as well, and attacker could add files to a repo that
would not be hashed but still checked out by kas. Avoid this by using an
empty .gitattributes in the worktree during the export in combination
with --worktree-attributes.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>