repos: Add caching support for sha256sum
To avoid going through a potentially length export step for each repo on
each kas invocation, save hash plus refspec used in a previous
successful validation aside the repo. This file is consulted first, and
if hash and refspec still match the kas configuration, validation is
skipped. If a repo fetch was triggered, the cache is dropped.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>